fix a bug with tag stats endpoint
The previous set up was causing the middleware functions to be called
twice
